mvrnn:getParameters now works properly.
Biggest lesson learned: calling getParameters NEEDS to be done before
any of the MVRNN's are created. The reason is that creating the networks
creates a number of clones (of, for instance, U and V) which share
parameters. But calling getParameters changes around how the parameters
are stored, and thus if these clones are created before getParameters
is called, then the clones will not have their storage changed
appropriately, and the clones will be pointing at different storage
areas (have different storage objects inside their Tensor objects)
than the originals. This is especially a problem since we're sharing
gradBias and gradWeight params.